✨ feat : gps 코드 분리
