+++ core/utils/mainPage.js
@@ -7,4 +7,28 @@ const returnTitle = (titleTag, searchKey = "") => {
   return "🏕️ " + titleCase[titleTag];
 };
 
-export default returnTitle;
+function getLocation(setData) {
+  if (navigator.geolocation) {
+    // GPS를 지원하면
+    navigator.geolocation.getCurrentPosition(
+      function (position) {
+        setData({
+          lati: position.coords.latitude,
+          long: position.coords.longitude,
+        });
+      },
+      function (error) {
+        console.error(error);
+      },
+      {
+        enableHighAccuracy: false,
+        maximumAge: 0,
+        timeout: Infinity,
+      }
+    );
+  } else {
+    alert("GPS를 지원하지 않습니다");
+  }
+}
+
+export { returnTitle, getLocation };
